requirements_gpu.txt
facenet-pytorch==2.4.1
matplotlib==3.3.0
opencv-python==4.5.1.48
keras==2.4.3
tensorflow-gpu==2.4.1
flask==1.1.2
flask_sqlalchemy==2.5.1
tqdm==4.60.0
seaborn==0.11.1